requirements.txt
APScheduler==3.9.1.post1
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachelib==0.4.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
chinesecalendar==1.8.0
click==8.0.4
clickhouse-connect==0.5.11
clicknium==0.1.11
clr-loader==0.2.4
colorama==0.4.6
coloredlogs==15.0.1
ddddocr==1.4.7
Deprecated==1.2.13
Flask==2.0.2
Flask-Script==2.0.6
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
flatbuffers==22.12.6
greenlet==1.1.2
hijri-converter==2.3.1
holidays==0.24
humanfriendly==10.0
idna==3.4
importlib-metadata==4.9.0
importlib-resources==5.10.0
itsdangerous==2.0.1
Jinja2==3.1.1
jsonschema==4.17.3
korean-lunar-calendar==0.3.1
loguru==0.6.0
lxml==4.9.2
lz4==4.3.2
MarkupSafe==2.1.1
mpmath==1.2.1
numpy==1.21.6
onnxruntime==1.13.1
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.5.3
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
protobuf==4.21.12
psycopg2-binary==2.9.3
pycparser==2.21
pycryptodome==3.17
PyExecJS==1.5.1
PyGithub==1.57
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.7
pyperclip==1.8.2
pyreadline3==3.4.1
pyrsistent==0.19.2
python-dateutil==2.8.2
pythonnet==3.0.1
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
redis==4.2.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.32
sympy==1.10.1
typing_extensions==4.0.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
Werkzeug==2.0.2
win32-setctime==1.1.0
wrapt==1.13.3
WTForms==3.0.0
zipp==3.6.0
zstandard==0.19.0
